Implemented new update and delete method in InsurancePolicyService.

+++ src/main/java/com/insurance/crm/service/impl/InsurancePolicyServiceImpl.java
@@ -1,8 +1,10 @@
 package com.insurance.crm.service.impl;
 
+import com.insurance.crm.constant.ErrorMessage;
 import com.insurance.crm.dto.policy.InsurancePolicyUpdateDto;
 import com.insurance.crm.entity.InsurancePolicy;
-import com.insurance.crm.exception.BadIdException;
+import com.insurance.crm.exception.NotDeletedException;
+import com.insurance.crm.exception.NotFoundException;
 import com.insurance.crm.repository.AgentRepository;
 import com.insurance.crm.repository.FiliationRepository;
 import com.insurance.crm.repository.InsurancePolicyRepository;
@@ -14,8 +16,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-
-import static com.insurance.crm.constant.ErrorMessage.INSURANCE_POLICY_NOT_FOUND_BY_ID;
 
 @Service
 @AllArgsConstructor
@@ -38,17 +38,22 @@ public class InsurancePolicyServiceImpl implements InsurancePolicyService {
 
     @Override
     public InsurancePolicy update(InsurancePolicyUpdateDto dto, Long id) {
-        InsurancePolicy insurancePolicy = insurancePolicyRepository.findById(id)
-                .orElseThrow(()-> new BadIdException(INSURANCE_POLICY_NOT_FOUND_BY_ID + id));
-        insurancePolicy.setFiliation(dto.getFiliation());
-        insurancePolicy.setAgent(dto.getAgent());
-        return insurancePolicyRepository.save(insurancePolicy);
+        return insurancePolicyRepository.findById(id)
+                .map(insurancePolicy -> {
+                    insurancePolicy.setFiliation(filiationRepository.findById(dto.getFiliation().getId()).get());
+                    insurancePolicy.setAgent(agentRepository.findById(dto.getAgent().getId()).get());
+                    return insurancePolicyRepository.save(insurancePolicy);
+                })
+          .orElseThrow(()-> new NotFoundException(ErrorMessage.INSURANCE_POLICY_NOT_FOUND_BY_ID + id));
     }
 
 
 
     @Override
     public void delete(Long id) {
+        if(!(insurancePolicyRepository.findById(id).isPresent())){
+            throw new NotDeletedException(ErrorMessage.INSURANCE_POLICY_NOT_DELETED + id);
+        }
         insurancePolicyRepository.deleteById(id);
     }
